meta init with full model run
