+++ fast_llm/models/auto.py
@@ -3,6 +3,7 @@ from fast_llm.engine.training.config import TrainerConfig
 from fast_llm.models.custom.config import CustomModelConfig, CustomTrainerConfig
 from fast_llm.models.gpt.config import GPTModelConfig, GPTTrainerConfig
 from fast_llm.utils import Registry
+from fast_llm.models.ssm.config import HybridTrainerConfig
 
 model_registry = Registry[str, FastLLMModelConfig](
     "Model",
@@ -22,6 +23,7 @@ trainer_registry = Registry[str, TrainerConfig](
         for trainer in [
             GPTTrainerConfig,
             CustomTrainerConfig,
+            HybridTrainerConfig,
         ]
     },
 )
